requirements.txt
streamlit
opencv-python
insightface
onnxruntime
faiss-cpu
pillow
numpy
scikit-learn
watchdog

